Change data type of price from float to big decimal.

+++ src/main/kotlin/com/noorifytech/maha/core/domain/Product.kt
@@ -1,9 +1,11 @@
 package com.noorifytech.maha.core.domain
 
+import java.math.BigDecimal
+
 data class Product(
         val id: String,
         val name: String,
-        val unitPrice: Float,
-        val discountPrice: Float = unitPrice,
+        val unitPrice: BigDecimal,
+        val discountPrice: BigDecimal = unitPrice,
         val discountQty: Int = 1
 )
